split up machinelearning_helpers
