+++ smart_importer/machinelearning_helpers.py
@@ -1,10 +1,9 @@
 """Machine Learning Helpers."""
 
-import json
 import logging
 from typing import List, Union, Tuple, NamedTuple
 
-import numpy as np
+import numpy
 from beancount import loader
 from beancount.core.data import Transaction, Posting, TxnPosting, filter_txns
 from sklearn.base import BaseEstimator, TransformerMixin
@@ -49,70 +48,6 @@ def load_training_data(
     return training_data
 
 
-def add_posting_to_transaction(transaction: Transaction,
-                               account: str) -> Transaction:
-    """Adds an empty posting with the given account to a transaction."""
-
-    if len(transaction.postings) != 1:
-        return transaction
-
-    additional_posting: Posting
-    additional_posting = Posting(account, None, None, None, None, None)
-    transaction.postings.append(additional_posting)
-    return transaction
-
-
-def add_payee_to_transaction(transaction: Transaction,
-                             payee: str,
-                             overwrite=False) -> Transaction:
-    """Sets a transactions's payee."""
-    if not transaction.payee or overwrite:
-        transaction = transaction._replace(payee=payee)
-    return transaction
-
-
-METADATA_KEY_SUGGESTED_ACCOUNTS = '__suggested_accounts__'
-METADATA_KEY_SUGGESTED_PAYEES = '__suggested_payees__'
-
-
-def add_suggested_accounts_to_transaction(
-        transaction: Transaction, suggestions: List[str]) -> Transaction:
-    """Adds suggested related accounts to a transaction."""
-    return _add_suggestions_to_transaction(
-        transaction, suggestions, key=METADATA_KEY_SUGGESTED_ACCOUNTS)
-
-
-def add_suggested_payees_to_transaction(transaction: Transaction,
-                                        suggestions: List[str]) -> Transaction:
-    """Adds suggested payees to a transaction."""
-    return _add_suggestions_to_transaction(
-        transaction, suggestions, key=METADATA_KEY_SUGGESTED_PAYEES)
-
-
-def _add_suggestions_to_transaction(transaction: Transaction,
-                                    suggestions: List[str],
-                                    key='__suggestions__'):
-    """
-    Adds a list of suggestions to a transaction under transaction.meta[key].
-    """
-    meta = transaction.meta
-    meta[key] = json.dumps(suggestions)
-    transaction = transaction._replace(meta=meta)
-    return transaction
-
-
-def merge_non_transaction_entries(imported_entries, enhanced_transactions):
-    enhanced_entries = []
-    enhanced_transactions_iter = iter(enhanced_transactions)
-    for entry in imported_entries:
-        if isinstance(entry, Transaction):
-            enhanced_entries.append(next(enhanced_transactions_iter))
-        else:
-            enhanced_entries.append(entry)
-
-    return enhanced_entries
-
-
 TxnPostingAccount = NamedTuple('TxnPostingAccount',
                                [('txn', Transaction), ('posting', Posting),
                                 ('account', str)])
@@ -127,7 +62,7 @@ class ArrayCaster(BaseEstimator, TransformerMixin):
         return self
 
     def transform(self, data):
-        return np.transpose(np.matrix(data))
+        return numpy.transpose(numpy.matrix(data))
 
 
 class NoFitMixin:
